Updated the top menu with a better UI and a home button
