+++ resources/META-INF/plugin.xml
@@ -10,6 +10,7 @@
 
   <change-notes><![CDATA[
       <ul>
+        <li>Added a new section at the top of the tool window where users can go to their my page, refresh and re-authenticate</li>
         <li>Plug-in will remember the last opened artifact through refreshes and restarts</li>
         <li>Added a SpiraTeam section in the "Tools" menu</li>
         <li>Fixed conflicting component names bug</li>
@@ -36,18 +37,24 @@
 
   <actions>
     <group id="SpiraTeam.ActiveToolbar" text="SpiraTeam" popup="true">
+      <!-- Need to reference the actions for them to show up -->
       <reference id="SpiraTeamLogin"/>
       <reference id="SpiraTeamRefresh"/>
+      <reference id="SpiraTeamMyPage"/>
       <add-to-group group-id="ToolsMenu" anchor="first"/>
     </group>
 
     <!-- Add your actions here -->
-    <action id="SpiraTeamLogin" class="com.inflectra.idea.ui.SpiraTeamLogin" text="SpiraTeam Login"
+    <action id="SpiraTeamLogin" class="com.inflectra.idea.ui.SpiraTeamLogin" text="Login"
             description="Prompt to login to SpiraTeam">
+      <add-to-group group-id="SpiraTeam.ActiveToolbar" anchor="last"/>
+    </action>
+    <action id="SpiraTeamRefresh" class="com.inflectra.idea.ui.SpiraTeamRefresh" text="Refresh"
+            description="Refresh the artifacts from SpiraTeam SpiraTeam">
       <add-to-group group-id="SpiraTeam.ActiveToolbar" anchor="first"/>
     </action>
-    <action id="SpiraTeamRefresh" class="com.inflectra.idea.ui.SpiraTeamRefresh" text="SpiraTeam Refresh"
-            description="Refresh the artifacts from SpiraTeam SpiraTeam">
+    <action id="SpiraTeamMyPage" class="com.inflectra.idea.ui.SpiraTeamGotoMyPage" text="My Page"
+            description="Open the SpiraTeam home page in browser">
       <add-to-group group-id="SpiraTeam.ActiveToolbar" anchor="last"/>
     </action>
   </actions>
